requirements.txt
langchain
langgraph
streamlit
chromadb
openai
langchain-community
langchain-openai
unstructured
numpy>=2.0.0,<2.3.0
protobuf>=6.30.0,<7.0.0
pillow>=11.0.0,<11.3.0
opentelemetry-proto==1.34.1
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai
streamlit
pandas
plotly
googlesearch-python
pycountry
langchain_openai
langchain_community
langchain
fuzzywuzzy
ollama
pgvector